Camera can pan
Issues may arise later, if so bring up an issue on gitHub. The script works by checking if mouse middle is down. If it is it checks the difference between the original position of the mouse and the current position. It moves the camera by this amount.

dragSpeed is a public variable that can be controlled in the unity editor. Changing the value changes the drag speed.